requirements.txt
mcp>=1.26.0
openai>=2.24.0
chromadb>=1.5.2
networkx>=3.6.1
python-dotenv>=1.2.1
pyyaml>=6.0
pyvis>=0.3.2
